Mekadia website source code
The source code for the main Mekadia.com website. The website combines
the functionality of a course management system, along with the common
features found in a social networking site, and the advanced question
and answer features found on Quora. Unfortunately some of the code has
been lost and is unrecoverable.